refactor(#1116): split cqlite-flight/src/agg.rs's unit tests into agg_tests.rs
`agg.rs` was 870 lines, over the campsite source target of 800, so the gate's
file-size ratchet FAILs any growth of it — and #3935's fix to the Min/Max
comparator needs both a code change there and a regression pin. Moved the inline
`#[cfg(test)] mod tests` body verbatim (de-indented one level) into a sibling
`agg_tests.rs`, wired with this crate's established
`#[cfg(test)] #[path = "..."] mod tests;` convention (filter.rs:663,
batch_bytes.rs:448, admission.rs:688).

Deliberately a `--lib` unit module and NOT a target under `cqlite-flight/tests/`:
`flight-tests` executes this crate's `--lib --bins` only and names the 42
integration targets it does not run, so a pin there would execute nowhere
(#3522/#3384).

agg.rs 870 -> 695 lines; agg_tests.rs 186. Same 18 agg-matching lib tests pass.
